(Alynna) Added ARRAY_INTERPRET.  If you know $lib/alynna, it performs exactly like str .. cat.  If you dont, it performs exactly like array_join, except dbrefs are translated to names, and it takes only the array to be interpreted, so its format is ( a -- s ) not ( a s1 -- s2 ) like array_join.
Also: Added STR, CAT, and }cat:
str is a define for {
cat is a define for } array_make array_interpret
}cat is a define for } array_make array_interpret

This allows code like this:
 me @ { me @ " has " 2 3 + " tails." }cat notify

If you use $lib/alynna, please get the newest version off of
http://frogpond.dyndns.org, which will undefine its str .. cat if it can
already find it.
This should make prograns that use these functions MUCH faster.