(Alynna) Modified ARRAY_INTERPRET to spit out 15 digit precision, truncated floats.  Therefore { 2.0 3.0 + }cat = "5" and { 2.1 3.2 + }cat = "5.3".
Added FTOSTRC ( f -- s ), translates a single float into a string using the
method above.  FTOSTR remains to allow you to get the extra digit of
precision kept internally, and so you have a float with extra zeroes for
routines that like to round using strcut pop.

Added the inserver defines
}cat = } array_make array_interpret

str = {      ($lib/alynna compatibility)
cat = }cat   ($lib/alynna compatibility)

If you use $lib/alynna, and upgrade to this, get the newest version from the
repository, or remove the str and cat _defs/ on the library.  It should cause
as significant increase, and not cause muftops to report libAlynna executing
a million times inna day or something.