New setup script.  Activates after make install, and it will set everything up nice and neat.  It also generates copies of your config and a restart script that actually works on something other than BSD.
This script is able to set up databases from FB6, and Glow, by setting the
proper tunes for maximum compatibility, and offers the option of installing
either minimal.proto or downloading and installing the protomuck basedb
developed by Moose.

It also rewrites parmfile.cfg, setting port numbers as you specify and other
pre-activation tunes.

For Winproto, I am planning on developing the equivalient of this script as
a Visual BASIC app, very soon.