Backported FB6's migration from GNU regex (regex.h) to PCRE (pcre.h). Ironically, the manpages for regexp and regsub had been backported from FB6 *after* they had converted away from GNU regex, so our manpages have been wrong and the REG_EXTENDED/4 flag has never done anything.
While backporting these changes I re-implemented as many of foxsteve's compiler
fixes as possible, but it may need more work.

There is an outstanding issue where the list returns from regexp are being
returned in the opposite order of what the manpage documents. It was like that
before these changes were comitted. I'll either fix this or update the manpage
sometime in the next week.