Make checkbashisms (mostly) happy; indicate phony targets in Makefile.in
Checkbashisms still complains about the use of $HOSTNAME, but since the setup
script explicitly defines $HOSTNAME to be the current hostname, it should work
regardless.

After installation, the install process instructs the user to use 'make setup',
which merely says 'setup is up to date'. Declare setup (and a few other non-
creation targets) phony so that make does not consider them fulfilled if a file
of the same name happens to exist.